requirements.txt
openai>=2.0
httpx>=0.28
python-telegram-bot>=22.0
slack-bolt>=1.27
slack-sdk>=3.40
irc>=20.0
python-dotenv>=1.0
feedparser>=6.0
requests>=2.32
pyyaml>=6.0
firecrawl-py>=4.0
